refactor(python): Split file of lazyframe methods
