+++ crates/polars-python/src/lazyframe/mod.rs
@@ -1,6 +1,7 @@
 mod exitable;
 #[cfg(feature = "pymethods")]
 mod general;
+mod io;
 #[cfg(feature = "pymethods")]
 mod serde;
 pub mod visit;
